requirements.txt
flask
flask-cors
gunicorn
tensorflow
numpy
pandas
scikit-learn
python-dotenv
twilio
googlemaps
